Added authentication for REST API

+++ app/api/views.py
@@ -4,8 +4,27 @@ from app.database import db
 from werkzeug.http import HTTP_STATUS_CODES
 from flask_login import current_user, login_required
 
+from flask import g
+from flask_httpauth import HTTPBasicAuth
+from app.models import User
+
 
 bp = Blueprint('api', __name__, url_prefix ='/api')
+
+basic_auth = HTTPBasicAuth()
+
+@basic_auth.verify_password
+def verify_password(username, password):
+    user = User.query.filter_by(username=username).first()
+    if user is None:
+        return False
+    g.current_user = user
+    return user.check_password(password)
+
+
+@basic_auth.error_handler
+def basic_auth_error():
+    return bad_request(401)
 
 
 def bad_request(status_code=400, message=None):
@@ -46,6 +65,7 @@ def get_posts():
 
 
 @bp.route('/posts', methods=['POST'])
+@basic_auth.login_required
 def create_post():
     data = request.get_json() or {}
     if ('name' not in data) or ('content' not in data):
@@ -53,8 +73,7 @@ def create_post():
         message = 'must include name and content fields'
         return bad_request(status_code, message)
     post = Post()
-    #post.user_id = current_user.id
-    post.user_id = 1 # for ApiPostCase in tests.py
+    post.user_id = g.current_user.id
     post.name = data['name']
     post.content = data['content']
     db.session.add(post)
@@ -71,13 +90,19 @@ def create_post():
     response.headers['Location'] = url_for('api.get_post', id=post.id)
     return response   
 
+
 @bp.route('/posts/<int:id>', methods=['PUT'])
+@basic_auth.login_required
 def update_post(id):
     post = Post.query.get_or_404(id)
     data = request.get_json() or {}
     if ('name' not in data) or ('content' not in data):
         status_code = 400
         message = 'must include name and content fields'
+        return bad_request(status_code, message)
+    if post.user_id != g.current_user.id:
+        status_code = 403
+        message = 'Permission error!'
         return bad_request(status_code, message)
     post.name = data['name']
     post.content = data['content']
@@ -96,8 +121,13 @@ def update_post(id):
 
 
 @bp.route('/posts/<int:id>', methods=['DELETE'])
+@basic_auth.login_required
 def delete_post(id):
     post = Post.query.get_or_404(id)
+    if post.user_id != g.current_user.id:
+        status_code = 403
+        message = 'Permission error!'
+        return bad_request(status_code, message)
     db.session.delete(post)
     db.session.commit()
     posts = Post.query.all()
